requirements.txt
# Python 3.11 Environment Dependencies
tensorflow>=2.19.0
fastapi>=0.115.0
uvicorn>=0.35.0
pillow>=11.3.0
numpy>=2.1.0
opencv-python>=4.11.0
matplotlib>=3.10.0
scikit-learn>=1.7.0
tqdm>=4.67.0
